requirements.txt
requests
beautifulsoup4
Pillow
rapidocr-onnxruntime
opencv-python-headless
numpy
pyperclip
